requirements.txt
tensorflow
google-cloud-aiplatform
tensorflow_hub
#tensorflow_text
openai
python-dotenv
oauth2client
vectorai